src/main/java/com/seoul/feedback/service/UserService.java
package com.seoul.feedback.service;

import com.seoul.feedback.dto.request.UserCreateRequest;
import com.seoul.feedback.entity.User;
import com.seoul.feedback.repository.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class UserService {

    private UserRepository userRepository;

    public User save(UserCreateRequest request) {
        return userRepository.save(User.builder()
                .login(request.getLogin())
                .build());
    }
}
